+++ src/discljord/messaging/specs.clj
@@ -238,7 +238,7 @@
 
 (s/def :command.option/type (set (vals command-option-types)))
 
-(s/def :command.option/name (string-spec #"[\w-]{1,32}"))
+(s/def :command.option/name (string-spec #"\S{1,32}"))
 (s/def :command.option/description (string-spec 1 100))
 (s/def :command.option/default boolean?)
 (s/def :command.option/required boolean?)
@@ -281,6 +281,15 @@
 
 (s/def :discljord.messaging.specs.command/permissions (s/coll-of :command/permission))
 
+(s/def :command/id ::command-id)
+
+(s/def :discljord.messaging.specs.command.guild/permissions
+  (s/keys :req-un [:command/id
+                   :discljord.messaging.specs.command/permissions]))
+
+(s/def :discljord.messaging.specs.command.guild/permissions-array
+  (s/coll-of :discljord.messaging.specs.command.guild/permissions))
+
 (s/def :discljord.messaging.specs.command/options
   (s/and (s/coll-of :command/option)
          (comp (partial >= 25) count)
@@ -294,7 +303,7 @@
 
 (s/def :command.option/options :discljord.messaging.specs.command/options)
 
-(s/def :discljord.messaging.specs.command/name (string-spec #"[\w-]{1,32}"))
+(s/def :discljord.messaging.specs.command/name (string-spec #"\S{1,32}"))
 
 (s/def :discljord.messaging.specs.command/description (string-spec 1 100))